chore: rename index files to component name
