+++ src/sections/create-collection/CreateCollection.tsx
@@ -6,7 +6,7 @@ import { useLoading } from '../loading/LoadingContext'
 import { useSession } from '../session/SessionContext'
 import { RequiredFieldText } from '../shared/form/RequiredFieldText/RequiredFieldText'
 import { BreadcrumbsGenerator } from '../shared/hierarchy/BreadcrumbsGenerator'
-import { CollectionForm, CollectionFormData } from './collection-form'
+import { CollectionForm, CollectionFormData } from './collection-form/CollectionForm'
 import { SeparationLine } from '../shared/layout/SeparationLine/SeparationLine'
 import { PageNotFound } from '../page-not-found/PageNotFound'
 import { CreateCollectionSkeleton } from './CreateCollectionSkeleton'
