+++ src/sections/create-collection/collection-form/metadata-fields-section/MetadataFieldsSection.tsx
