+++ src/sections/create-collection/collection-form/top-fields-section/TopFieldsSection.tsx
